requirements.txt
crewai
langchain
langchain-community
streamlit
duckduckgo-search
python-dotenv
tavily-python
langchain-google-genai
